% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addmessage.R
\name{addmessage}
\alias{addmessage}
\title{Putting a New Chat Message into Old Message(s).}
\usage{
addmessage(oldmessage, newmessage)
}
\arguments{
\item{oldmessage}{Old message or message list.}

\item{newmessage}{New message.}
}
\value{
Merged chat message list.
}
\description{
Conveniently constructs chat message list to be passed as arguments to the chatgpt API.
}
\examples{
message_A <- list(role = "user", content = "Hello!")
message_B <- list(role = "assistant", content = "Hello!")
message_C <- usermessage("How are you?")
message_step1 <- addmessage(message_A, message_B)
message_step2 <- addmessage(message_step1, message_C)

}
