% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistantmessage.R
\name{assistantmessage}
\alias{assistantmessage}
\title{Create an Assistant Message.}
\usage{
assistantmessage(message)
}
\arguments{
\item{message}{Message text string.}
}
\value{
A two-element list consisting of role = "assistant", content = message.
}
\description{
Create an Assistant Message.
}
\examples{
assistantmessage_1 <- assistantmessage("How can I help you?")

}
