% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prompttemp.R
\name{check_prompttemp}
\alias{check_prompttemp}
\title{Check the Prompt Template File.}
\usage{
check_prompttemp(temptype, oldprompttemp)
}
\arguments{
\item{temptype}{Type string of the template.}

\item{oldprompttemp}{Built-in template}
}
\value{
prompt template text string
}
\description{
Check if the prompt template file exists, if it does, import it, if it doesn't, create a .md file and write the built-in template, then import it.
}
\examples{
# It is best not to call this function independently.

}
