% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Coalesce operator (re-export from rlang)
See \link[rlang:op-null-default]{rlang::\%||\%}.}
\description{
Coalesce operator (re-export from rlang)
See \link[rlang:op-null-default]{rlang::\%||\%}.
}
