% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coalesce.R
\name{coalesce_len0}
\alias{coalesce_len0}
\title{Coalesce NULL or length-0 vectors to a default}
\usage{
coalesce_len0(x, default)
}
\arguments{
\item{x}{Any object.}

\item{default}{Value to return when x is NULL or length(x) == 0.}
}
\value{
Either x or default.
}
\description{
Coalesce NULL or length-0 vectors to a default
}
\keyword{internal}
