% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_to_text.R
\name{dataframe_to_text}
\alias{dataframe_to_text}
\title{Convert data.frams to text interpretable by ChatGPT}
\usage{
dataframe_to_text(dataframe)
}
\arguments{
\item{dataframe}{A dataframe to be converted.}
}
\value{
A character vector representing the dataframe in text format.
}
\description{
This function takes a dataframe as input and converts it to a text representation.
}
\examples{
dataframe <- data.frame(
  Name = c("John", "Alice", "Bob"),
  Age = c(25, 30, 22),
  Score = c(95, 88, 75)
)
dataframe_to_text(dataframe)

}
