% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_cache.R
\name{delete_models_cache}
\alias{delete_models_cache}
\title{Clear models cache so that the next run will re-probe the server.}
\usage{
delete_models_cache()
}
\value{
invisible TRUE
}
\description{
Clear models cache so that the next run will re-probe the server.
}
