% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_cache.R
\name{.cache_put}
\alias{.cache_put}
\title{Save a cache entry for provider+base_url with a vector of model IDs and timestamp.}
\usage{
.cache_put(provider, base_url, models, base_url_normalized = FALSE)
}
\arguments{
\item{base_url_normalized}{Logical; set TRUE when \code{base_url} is already normalized.}
}
\description{
Save a cache entry for provider+base_url with a vector of model IDs and timestamp.
}
\keyword{internal}
