% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.in_allowed}
\alias{.in_allowed}
\title{check if a scalar belongs to an allowed set (case-insensitive for character)}
\usage{
.in_allowed(value, allowed)
}
\description{
check if a scalar belongs to an allowed set (case-insensitive for character)
}
\keyword{internal}
