% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_na_like}
\alias{.is_na_like}
\title{Determine if a scalar should be treated as NA (case-insensitive for characters)}
\usage{
.is_na_like(x, na_vals = c("NA", "null", "", "[]", "{}", "None"))
}
\description{
Determine if a scalar should be treated as NA (case-insensitive for characters)
}
\keyword{internal}
