% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.parse_key_spec}
\alias{.parse_key_spec}
\title{Interpret a key spec (type string or allowed set)}
\usage{
.parse_key_spec(spec)
}
\arguments{
\item{spec}{Either a type string ("integer","numeric","character","logical")
or a vector of allowed values.}
}
\value{
list(type = <chr|NULL>, allowed = <vector|NULL>)
}
\description{
Interpret a key spec (type string or allowed set)
}
\keyword{internal}
