% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.row_df}
\alias{.row_df}
\title{list_models() helprer}
\usage{
.row_df(
  provider,
  base_url,
  models_df,
  availability,
  src,
  ts,
  status = NA_character_
)
}
\description{
list_models() helprer
}
\keyword{internal}
