% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_fix_parse_validate.R
\name{.tidy_json}
\alias{.tidy_json}
\title{Clean + repair model JSON output in a safe, staged way.
Returns a list(txt = \if{html}{\out{<string>}}, log = \if{html}{\out{<character vector of applied fixes>}})}
\usage{
.tidy_json(
  x,
  na_values = c("NA", "null", "", "[]", "{}", "None"),
  aggressive = FALSE
)
}
\description{
Clean + repair model JSON output in a safe, staged way.
Returns a list(txt = \if{html}{\out{<string>}}, log = \if{html}{\out{<character vector of applied fixes>}})
}
\keyword{internal}
