% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{extract_code}
\alias{extract_code}
\title{Extract Code String from JSON Structure String.}
\usage{
extract_code(json_string)
}
\arguments{
\item{json_string}{JSON string to be extracted}
}
\value{
Code string.
}
\description{
Extracts code text content from the structured JSON string returned by the model.
}
\examples{
Jsonstring <- '{"code":"x=1"}'
codestring <- extract_code(Jsonstring)

}
