% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-object.R
\name{format.gptr_session}
\alias{format.gptr_session}
\alias{as.character.gptr_session}
\title{Format a session as its text}
\usage{
\method{format}{gptr_session}(x, ...)

\method{as.character}{gptr_session}(x, ...)
}
\arguments{
\item{x}{A \code{gptr_session}.}

\item{...}{Unused.}
}
\value{
\code{x$text}.
}
\description{
Format a session as its text
}
