% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_prompttemp.R
\name{give_prompttemp}
\alias{give_prompttemp}
\title{Provide Prompt Templates.}
\usage{
give_prompttemp(temptype)
}
\arguments{
\item{temptype}{The type of the template.}
}
\value{
Prompt template text string.
}
\description{
There are three types of prompt templates: "coding", "debugging" and "asking".
With this function, you can see the current content of the three types of prompt templates.
}
\examples{
give_prompttemp("coding")

}
