% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list.R
\name{model_list}
\alias{model_list}
\title{Get a List of Available Models (Only for Openai Format).}
\usage{
model_list()
}
\value{
Models list.
}
\description{
Get a List of Available Models (Only for Openai Format).
}
\examples{
\donttest{
modellist <- model_list()
}

}
