% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R
\name{model_table}
\alias{model_table}
\title{Get a Data Frame of Available Models(Only for Openai Format).}
\usage{
model_table()
}
\value{
Models data frame.
}
\description{
Get a Data Frame of Available Models(Only for Openai Format).
}
\examples{

model_table()

}
