% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{play}
\alias{play}
\title{Play a gptr_audio}
\usage{
play(x)
}
\arguments{
\item{x}{A \code{gptr_audio} object (as returned by \code{\link[=gpt_tts]{gpt_tts()}}), or a path string.}
}
\description{
Play a gptr_audio
}
