% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_tts.R
\name{print.gptr_audio}
\alias{print.gptr_audio}
\title{Print method for gptr audio objects}
\usage{
\method{print}{gptr_audio}(x, ...)
}
\arguments{
\item{x}{A \code{gptr_audio} object returned by \code{\link[=gpt_tts]{gpt_tts()}}.}

\item{...}{Unused; present for method compatibility.}
}
\description{
Print method for gptr audio objects
}
