% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-object.R
\name{print.gptr_session_summary}
\alias{print.gptr_session_summary}
\title{Print a session summary}
\usage{
\method{print}{gptr_session_summary}(x, ...)
}
\arguments{
\item{x}{A \code{gptr_session_summary}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a session summary
}
