% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prompttemp.R
\name{read_prompttemp}
\alias{read_prompttemp}
\title{Importing Prompt Template from .md File.}
\usage{
read_prompttemp(filename)
}
\arguments{
\item{filename}{File name of the .md file.}
}
\value{
Prompt template text string.
}
\description{
Importing Prompt Template from .md File.
}
\examples{
\donttest{
prompttemp <- read_prompttemp("temp.md")
}

}
