% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_chat.R
\name{start_chat}
\alias{start_chat}
\title{Select a Particular API Interface Function for a Single Chat.}
\usage{
start_chat(message, model, showtokennum = TRUE, chat_function = NULL)
}
\arguments{
\item{message}{Text string passed directly to the model.}

\item{model}{The name string of the model you want to use.}

\item{showtokennum}{Whether to show the token's usage, defaults to TRUE.}

\item{chat_function}{A custom model chat API interface function that takes three parameters, message, model, and tokennum, and returns a list consisting of answer and status_code. Defaults to NULL, at which point the built-in openai style interface function is called. For custom function styles, see the function mychat().}
}
\value{
A two-element list consisting of answer(answer of the model) and status_code(Internet access status code, 200 for subsequent functions to work properly).
}
\description{
Select a Particular API Interface Function for a Single Chat.
}
\examples{
\donttest{
# Use the openai interface function by default

New_chat <- start_chat("Hello!","gpt-4",chat_function=mychat)
cat(New_chat$answer)

}
}
