% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemmessage.R
\name{systemmessage}
\alias{systemmessage}
\title{Create a System Message.}
\usage{
systemmessage(message)
}
\arguments{
\item{message}{Message text string.}
}
\value{
A two-element list consisting of role = "system", content = message.
}
\description{
Create a System Message.
}
\examples{
systemmessage_1 <- systemmessage("Hello!")

}
