% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_text.R
\name{trim_text}
\alias{trim_text}
\title{Trim text by "tokens" without hard Python deps}
\usage{
trim_text(
  text,
  max_tokens = Inf,
  token_mode = c("words", "chars", "custom"),
  custom_tokenizer = NULL
)
}
\arguments{
\item{text}{character scalar}

\item{max_tokens}{integer max boundary}

\item{token_mode}{"words" (default), "chars", or "custom"}

\item{custom_tokenizer}{optional function(text, max_tokens) -> trimmed_text}
}
\description{
Trim text by "tokens" without hard Python deps
}
