% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usermessage.R
\name{usermessage}
\alias{usermessage}
\title{Create a User Message.}
\usage{
usermessage(message)
}
\arguments{
\item{message}{Message text string.}
}
\value{
A two-element list consisting of role = "user", content = message.
}
\description{
Create a User Message.
}
\examples{
usermessage_1 <- usermessage("Hello!")

}
