% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_prompttemp.R
\name{write_prompttemp}
\alias{write_prompttemp}
\title{Exporting Prompt Template from .md File.}
\usage{
write_prompttemp(md_text, filename)
}
\arguments{
\item{md_text}{Prompt template text string.}

\item{filename}{File name of the .md file.}
}
\value{
Nothing.
}
\description{
Exporting Prompt Template from .md File.
}
\examples{
write_prompttemp("Template content.", "temp.md")

}
