% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{.as_models_df}
\alias{.as_models_df}
\title{list_models() helprer
internal
base_url_normalized Logical; set TRUE when base_url is already normalized}
\usage{
.as_models_df(x)
}
\description{
list_models() helprer
internal
base_url_normalized Logical; set TRUE when base_url is already normalized
}
