% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{.row_df}
\alias{.row_df}
\title{list_models() helper}
\usage{
.row_df(
  provider,
  base_url,
  models_df,
  availability,
  src,
  ts,
  status = NA_character_,
  base_url_normalized = FALSE
)
}
\description{
list_models() helper
}
\keyword{internal}
