% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_cache.R
\name{refresh_models}
\alias{refresh_models}
\title{Refresh model listings, bypassing cache}
\usage{
refresh_models(...)
}
\value{
See \code{\link[=list_models]{list_models()}}
}
\description{
Convenience wrapper around \code{\link[=list_models]{list_models()}} with \code{refresh = TRUE}.
}
